+++ datahub-graphql-core/src/main/java/com/linkedin/datahub/graphql/types/ermodelrelationship/CreateERModelRelationshipResolver.java
@@ -1,4 +1,4 @@
-package com.linkedin.datahub.graphql.types.ermodelrelation;
+package com.linkedin.datahub.graphql.types.ermodelrelationship;
 
 import static com.linkedin.datahub.graphql.resolvers.ResolverUtils.bindArgument;
 
@@ -11,8 +11,8 @@ import com.linkedin.datahub.graphql.exception.AuthorizationException;
 import com.linkedin.datahub.graphql.generated.ERModelRelationPropertiesInput;
 import com.linkedin.datahub.graphql.generated.ERModelRelationship;
 import com.linkedin.datahub.graphql.generated.ERModelRelationshipUpdateInput;
-import com.linkedin.datahub.graphql.types.ermodelrelation.mappers.ERModelRelationMapper;
-import com.linkedin.datahub.graphql.types.ermodelrelation.mappers.ERModelRelationshipUpdateInputMapper;
+import com.linkedin.datahub.graphql.types.ermodelrelationship.mappers.ERModelRelationMapper;
+import com.linkedin.datahub.graphql.types.ermodelrelationship.mappers.ERModelRelationshipUpdateInputMapper;
 import com.linkedin.entity.client.EntityClient;
 import com.linkedin.metadata.service.ERModelRelationshipService;
 import com.linkedin.mxe.MetadataChangeProposal;
