+++ datahub-graphql-core/src/main/java/com/linkedin/datahub/graphql/types/ermodelrelationship/UpdateERModelRelationshipResolver.java
@@ -1,4 +1,4 @@
-package com.linkedin.datahub.graphql.types.ermodelrelation;
+package com.linkedin.datahub.graphql.types.ermodelrelationship;
 
 import static com.linkedin.datahub.graphql.resolvers.ResolverUtils.bindArgument;
 
@@ -7,7 +7,7 @@ import com.linkedin.common.urn.ERModelRelationshipUrn;
 import com.linkedin.datahub.graphql.QueryContext;
 import com.linkedin.datahub.graphql.exception.AuthorizationException;
 import com.linkedin.datahub.graphql.generated.ERModelRelationshipUpdateInput;
-import com.linkedin.datahub.graphql.types.ermodelrelation.mappers.ERModelRelationshipUpdateInputMapper;
+import com.linkedin.datahub.graphql.types.ermodelrelationship.mappers.ERModelRelationshipUpdateInputMapper;
 import com.linkedin.entity.client.EntityClient;
 import com.linkedin.mxe.MetadataChangeProposal;
 import com.linkedin.r2.RemoteInvocationException;
